Revert "Prune modules and kselftests from images"
This reverts commit 95389e96ca9e91acb3b2b818c5bd35c3fef5849b.

It was a bad idea: Mounting the ext image and then removing
files from it caused it to expand fully; compressing it
afterwards had very little effect, and the images resulted
in gigantic 2+ GB images to download.

This was only effective for tarballs, but those were already
petite to begin with.

Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>